+++ changelog.rst
@@ -18,6 +18,8 @@ Changes by date
 **version 0.5.3**
 
 * (logic) debugged variance error calculations for Area-to-Point Poisson Kriging
+* (bug) https://github.com/DataverseLabs/pyinterpolate/issues/428
+* (enhancement) added universal kriging functionality along with multivariate regression
 
 2024-06-26
 ----------